Make iterate/app own its RPC adapter
