+++ apps/os/src/domains/workers/iterate-app-virtual-module.test.ts
@@ -7,7 +7,6 @@ test("the embedded iterate/app runtime is loader-ready plain JavaScript", async 
     esbuild.transform(ITERATE_APP_VIRTUAL_MODULE, { format: "esm", loader: "js" }),
   ).resolves.toBeDefined();
 
-  expect(ITERATE_APP_VIRTUAL_MODULE).toContain("newWorkersWebSocketRpcResponse");
   expect(ITERATE_APP_VIRTUAL_MODULE).toContain("LiveStateRpcTarget");
   expect(ITERATE_APP_VIRTUAL_MODULE).toContain('from "@iterate-com/capnweb"');
   expect(ITERATE_APP_VIRTUAL_MODULE).not.toContain("import type");
