+++ apps/os/src/domains/workers/iterate-live-state-virtual-module.generated.ts
@@ -1,0 +1,7 @@
+// The public `iterate/live-state` runtime bundled to loader-ready JavaScript
+// and embedded as a virtual module. Edit packages/iterate, then run
+// `pnpm lint:fix`; drift is a lint error. The codegen preset owns formatting.
+// codegen:start {preset: custom, source: ./iterate-live-state-virtual-module.codegen.cjs, export: iterateLiveStateVirtualModule}
+export const ITERATE_LIVE_STATE_VIRTUAL_MODULE =
+  '// packages/iterate/src/itx/live-state/diff.ts\nfunction diff(prev, next) {\n  if (Object.is(prev, next)) return void 0;\n  if (!isPlainObject(prev) || !isPlainObject(next)) return { set: next };\n  const fields = [];\n  const drop = [];\n  for (const key of Object.keys(next)) {\n    if (next[key] === void 0) {\n      if (Object.hasOwn(prev, key)) drop.push(key);\n      continue;\n    }\n    const childPatch = diff(Object.hasOwn(prev, key) ? prev[key] : void 0, next[key]);\n    if (childPatch !== void 0) fields.push([key, childPatch]);\n  }\n  for (const key of Object.keys(prev)) {\n    if (!Object.hasOwn(next, key)) drop.push(key);\n  }\n  if (fields.length === 0 && drop.length === 0) return void 0;\n  const patch = {};\n  if (fields.length > 0) patch.fields = Object.fromEntries(fields);\n  if (drop.length > 0) patch.drop = drop;\n  return patch;\n}\nfunction applyPatch(prev, patch) {\n  if ("set" in patch) return patch.set;\n  const base = isPlainObject(prev) ? prev : {};\n  const next = { ...base };\n  if (patch.fields) {\n    for (const [key, childPatch] of Object.entries(patch.fields)) {\n      Object.defineProperty(next, key, {\n        value: applyPatch(Object.hasOwn(base, key) ? base[key] : void 0, childPatch),\n        enumerable: true,\n        writable: true,\n        configurable: true\n      });\n    }\n  }\n  if (patch.drop) {\n    for (const key of patch.drop) delete next[key];\n  }\n  return next;\n}\nfunction isPlainObject(value) {\n  if (typeof value !== "object" || value === null) return false;\n  const proto = Object.getPrototypeOf(value);\n  return proto === Object.prototype || proto === null;\n}\n\n// packages/iterate/src/itx/live-state/store.ts\nfunction createLiveStateStore() {\n  let held = { revision: -1, state: void 0 };\n  const listeners = /* @__PURE__ */ new Set();\n  const notify = () => listeners.forEach((listener) => listener());\n  return {\n    getState: () => held.state,\n    subscribe: (listener) => {\n      listeners.add(listener);\n      return () => void listeners.delete(listener);\n    },\n    reset: () => {\n      held = { revision: -1, state: void 0 };\n      notify();\n    },\n    apply: (update, resync) => {\n      if (update.type === "snapshot") {\n        held = { revision: update.revision, state: update.state };\n      } else if (update.from !== held.revision) {\n        resync();\n        return;\n      } else {\n        held = { revision: update.to, state: applyPatch(held.state, update.patch) };\n      }\n      notify();\n    }\n  };\n}\n\n// packages/iterate/src/itx/rpc/retain.ts\nfunction retainCallback(callback) {\n  const retainable = callback;\n  const retained = retainable.dup?.() ?? retainable;\n  const dispose = retained[Symbol.dispose]?.bind(retained);\n  let disposed = false;\n  const wrapped = Object.assign((arg) => retained(arg), {\n    [Symbol.dispose]() {\n      if (disposed) return;\n      disposed = true;\n      dispose?.();\n    }\n  });\n  const onRpcBroken = retained.onRpcBroken;\n  if (typeof onRpcBroken === "function") {\n    wrapped.onRpcBroken = (handler) => {\n      try {\n        const result = onRpcBroken.call(retained, handler);\n        if (isThenable(result)) void Promise.resolve(result).catch(() => {\n        });\n      } catch {\n      }\n    };\n  }\n  return wrapped;\n}\nfunction isThenable(value) {\n  return value !== null && (typeof value === "object" || typeof value === "function") && typeof value.then === "function";\n}\nfunction disposeIgnoredRpcResult(result) {\n  if (result !== null && (typeof result === "object" || typeof result === "function") && Symbol.dispose in result) {\n    result[Symbol.dispose]();\n  }\n}\n\n// packages/iterate/src/itx/live-state/engine.ts\nvar DEFAULT_DEBOUNCE_MS = 100;\nvar LiveState = class {\n  #current;\n  /** The state every current subscriber has been brought to — the diff baseline. */\n  #broadcast;\n  #revision = 0;\n  #subscribers = /* @__PURE__ */ new Set();\n  #debounceMs;\n  #flushTimer;\n  constructor(initial, options = {}) {\n    this.#current = initial;\n    this.#broadcast = initial;\n    this.#debounceMs = options.debounceMs ?? DEFAULT_DEBOUNCE_MS;\n  }\n  /** The current state — reflects every `setState`/`assign`, even while dormant. */\n  getState() {\n    return this.#current;\n  }\n  /** Replace the state. Build a NEW value (don\'t mutate) so the diff stays cheap. */\n  setState(next) {\n    this.#current = typeof next === "function" ? next(this.#current) : next;\n    if (this.#subscribers.size > 0) this.#scheduleFlush();\n  }\n  /** Shallow-merge a partial into the state — sugar for `setState(p => ({ ...p, ...partial }))`. */\n  assign(partial) {\n    this.setState((prev) => ({ ...prev, ...partial }));\n  }\n  subscribe(sink) {\n    if (this.#subscribers.size === 0) this.#broadcast = this.#current;\n    const retained = retainCallback(sink);\n    this.#subscribers.add(retained);\n    retained.onRpcBroken?.(() => this.#drop(retained));\n    this.#deliver(retained, { type: "snapshot", revision: this.#revision, state: this.#broadcast });\n    return {\n      ping: () => this.#subscribers.has(retained),\n      unsubscribe: () => this.#drop(retained),\n      [Symbol.dispose]: () => this.#drop(retained)\n    };\n  }\n  #scheduleFlush() {\n    this.#flushTimer ??= setTimeout(() => {\n      this.#flushTimer = void 0;\n      this.#flush();\n    }, this.#debounceMs);\n  }\n  #flush() {\n    const patch = diff(this.#broadcast, this.#current);\n    if (patch === void 0) return;\n    const from = this.#revision;\n    const to = this.#revision += 1;\n    this.#broadcast = this.#current;\n    const update = { type: "patch", from, to, patch };\n    for (const subscriber of [...this.#subscribers]) this.#deliver(subscriber, update);\n  }\n  /**\n   * Fire-and-forget delivery that observes rejection. A dead stub rejects every\n   * push (and can throw synchronously), so any failure drops the subscriber —\n   * otherwise a corpse would keep its slot and its `ping()` would lie forever.\n   */\n  #deliver(subscriber, update) {\n    let result;\n    try {\n      result = subscriber(update);\n    } catch {\n      this.#drop(subscriber);\n      return;\n    }\n    if (isThenable(result)) {\n      void Promise.resolve(result).then(void 0, () => this.#drop(subscriber)).finally(() => disposeIgnoredRpcResult(result));\n      return;\n    }\n    disposeIgnoredRpcResult(result);\n  }\n  #drop(subscriber) {\n    if (!this.#subscribers.delete(subscriber)) return;\n    subscriber[Symbol.dispose]();\n    if (this.#subscribers.size === 0 && this.#flushTimer !== void 0) {\n      clearTimeout(this.#flushTimer);\n      this.#flushTimer = void 0;\n    }\n  }\n};\nexport {\n  LiveState,\n  applyPatch,\n  createLiveStateStore,\n  diff,\n  disposeIgnoredRpcResult,\n  isThenable,\n  retainCallback\n};\n';
+// codegen:end
