+++ apps/os/src/domains/workers/worker-loader.ts
@@ -11,6 +11,7 @@ import {
 } from "./artifact-store.ts";
 import { workerBuildKey, type ResolvedWorkerFileSource } from "./build-key.ts";
 import { ITERATE_APP_VIRTUAL_MODULE } from "./iterate-app-virtual-module.generated.ts";
+import { ITERATE_LIVE_STATE_VIRTUAL_MODULE } from "./iterate-live-state-virtual-module.generated.ts";
 import { ITERATE_PROCESSORS_CLOUDFLARE_VIRTUAL_MODULE } from "./iterate-processors-cloudflare-virtual-module.generated.ts";
 import { ITERATE_PROCESSORS_VIRTUAL_MODULE } from "./iterate-processors-virtual-module.generated.ts";
 import { ITERATE_SDK_VIRTUAL_MODULE } from "./iterate-sdk-virtual-module.generated.ts";
@@ -152,6 +153,7 @@ function withIterateVirtualModules(options: WorkerBuildOptions): WorkerBuildOpti
     ...options,
     virtualModules: {
       "iterate/app": ITERATE_APP_VIRTUAL_MODULE,
+      "iterate/live-state": ITERATE_LIVE_STATE_VIRTUAL_MODULE,
       "iterate/sdk": ITERATE_SDK_VIRTUAL_MODULE,
       "iterate/processors": ITERATE_PROCESSORS_VIRTUAL_MODULE,
       "iterate/processors/cloudflare": ITERATE_PROCESSORS_CLOUDFLARE_VIRTUAL_MODULE,
