+++ knip.ts
@@ -28,6 +28,9 @@ function makeOsCloudflareAppWorkspace(workerEnvShim: string): WorkspaceConfig {
       "scripts/itx.ts",
       // Operational smoke for the create-project -> onboarding-greeting path.
       "e2e/vitest/onboarding-smoke.ts",
+      // Seeded as a standalone worker entry outside apps/os/src. Tests import
+      // parts of it, but the deployed config-repo worker uses the whole file.
+      "config-repo-template/worker.ts",
       // Used by apps/streams-example-app through its `~` alias into apps/os
       // src; knip does not resolve that cross-workspace alias.
       "src/domains/streams/client-libraries/processors/browser-event-feed/contract.ts",
